+++ apps/ui/components/application/TimeSelector.tsx
@@ -12,7 +12,7 @@ import {
 import { breakpoints } from "common/src/common/style";
 import { filterNonNullable, fromMondayFirstUnsafe } from "common/src/helpers";
 import { WEEKDAYS } from "common/src/const";
-import { arrowDown, arrowUp } from "@/styles/util";
+import { arrowDown, arrowUp } from "@/styled/util";
 import { TimePreview } from "./TimePreview";
 import { type ApplicationPage2FormValues } from "./form";
 import { useFormContext } from "react-hook-form";
