+++ apps/ui/styled/util.ts
@@ -1,4 +1,3 @@
-import { breakpoints } from "common";
 import { fontMedium, fontRegular } from "common/src/common/typography";
 import { Button } from "hds-react";
 import Link from "next/link";
@@ -53,31 +52,5 @@ export const InlineStyledLink = styled(Link)`
     color: var(--color-black);
     text-decoration: underline;
     ${fontRegular};
-  }
-`;
-
-const CARD_COLUMN_SIZE = 390;
-
-export const ReservationPageWrapper = styled.div<{ $nRows?: number }>`
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-template-rows: ${({ $nRows }) => `repeat(${$nRows ?? 4}, auto)`};
-  grid-gap: var(--spacing-m);
-  justify-content: space-between;
-  margin: 0;
-  padding: 0;
-
-  @media (min-width: ${breakpoints.m}) {
-    grid-template-columns: 1fr ${CARD_COLUMN_SIZE}px;
-  }
-`;
-
-// Larger breakpoint for reservation unit page because Calendar takes more space.
-export const ReservationUnitPageWrapper = styled(ReservationPageWrapper)`
-  @media (min-width: ${breakpoints.m}) {
-    grid-template-columns: 1fr;
-  }
-  @media (min-width: ${breakpoints.l}) {
-    grid-template-columns: 2fr 1fr;
   }
 `;
