+++ apps/ui/test/test.gql.utils.ts
@@ -1,15 +1,20 @@
 import {
   ApplicantTypeChoice,
   type ApplicationFormFragment,
+  ApplicationRoundFieldsFragment,
+  ApplicationRoundStatusChoice,
   ApplicationStatusChoice,
   CreateApplicationDocument,
   type CreateApplicationMutationResult,
   type CreateApplicationMutationVariables,
   CurrentUserDocument,
   type CurrentUserQuery,
+  IsReservableFieldsFragment,
+  OptionsQuery,
   OrganizationTypeChoice,
   Priority,
   ReservationKind,
+  ReservationStartInterval,
   ReservationUnitOrderingChoices,
   SearchReservationUnitsDocument,
   type SearchReservationUnitsQuery,
@@ -20,11 +25,10 @@ import {
 } from "@/gql/gql-types";
 import { base64encode } from "common/src/helpers";
 import { addDays, addYears } from "date-fns";
-import { DocumentNode } from "graphql";
-import {
-  createMockReservationUnitType,
-  generateNameFragment,
-} from "./test.utils";
+import { type DocumentNode } from "graphql";
+import { type TFunction } from "i18next";
+import { getDurationOptions } from "@/modules/const";
+import { type ReservableMap, type RoundPeriod } from "@/modules/reservable";
 
 export type CreateGraphQLMockProps = {
   noUser?: boolean;
@@ -42,6 +46,7 @@ export type CreateGraphQLMocksReturn = Array<{
   };
   error?: Error | undefined;
 }>;
+
 // TODO parametrize the variables
 // we need at least the following:
 // - no user
@@ -226,6 +231,7 @@ function createSearchVariablesMock({
   } as const;
 }
 
+/// @param page which page is valid (page0 => nothing is valid), preview => it's sent
 function createMockApplicationSection({
   page = "page0",
 }: {
@@ -440,5 +446,161 @@ export function createMockTermsOfUse(): TermsOfUseFieldsFragment {
     textFi: "Yleiset käyttöehdot",
     textEn: "General terms of use",
     textSv: "Allmänna användningsvillkor",
+  };
+}
+
+// Option mocks
+// TODO should move to test.utils.ts
+// TODO improve mockT so that we pull through the duration value (not just the minutes / hours)
+export const mockT: TFunction = ((key: string) => key) as TFunction;
+export const mockDurationOptions = getDurationOptions(mockT);
+export const mockReservationPurposesOptions = Array.from(
+  { length: 5 },
+  (_, i) => ({
+    pk: i + 1,
+  })
+).map(({ pk }) => ({ value: pk, label: `Reservation Purpose ${pk}` }));
+export const mockAgeGroupOptions = Array.from(
+  { length: 5 },
+  (_, i) => i + 1
+).map((v) => ({
+  pk: v,
+  maximum: v,
+  minimum: v,
+}));
+
+export function createOptionQueryMock(): OptionsQuery {
+  return {
+    reservationPurposes: {
+      edges: mockReservationPurposesOptions
+        .map(({ value, label }) => ({
+          id: base64encode(`ReservationPurposeNode:${value}`),
+          pk: value,
+          nameFi: label,
+          nameSv: label,
+          nameEn: label,
+        }))
+        .map((node) => ({ node })),
+    },
+    ageGroups: {
+      edges: mockAgeGroupOptions
+        .map(({ pk, maximum, minimum }) => ({
+          id: base64encode(`ReservationPurposeNode:${pk}`),
+          pk,
+          minimum,
+          maximum,
+        }))
+        .map((node) => ({ node })),
+    },
+    // NOT defining these atm because Seasonal Form doesn't use them
+    // but the query requires matching keys
+    reservationUnitTypes: {
+      edges: [],
+    },
+    purposes: {
+      edges: [],
+    },
+    cities: {
+      edges: [],
+    },
+    equipmentsAll: [],
+    unitsAll: [],
+  };
+}
+
+type ReservationUnitType = Omit<
+  IsReservableFieldsFragment,
+  "reservableTimeSpans"
+>;
+type MockReservationUnitProps = {
+  bufferTimeBefore?: number;
+  bufferTimeAfter?: number;
+  reservableTimes?: ReservableMap;
+  interval?: ReservationStartInterval;
+  maxReservationDuration?: IsReservableFieldsFragment["maxReservationDuration"];
+  minReservationDuration?: IsReservableFieldsFragment["minReservationDuration"];
+  activeApplicationRounds?: RoundPeriod[];
+  reservationsMinDaysBefore?: number;
+  reservationsMaxDaysBefore?: number | null;
+};
+/// create a mock for IsReservableFragment (not a full reservation unit)
+export function createMockReservationUnit({
+  bufferTimeBefore = 0,
+  bufferTimeAfter = 0,
+  interval = ReservationStartInterval.Interval_15Mins,
+  maxReservationDuration = 0,
+  minReservationDuration = 0,
+  reservationsMinDaysBefore = 0,
+  reservationsMaxDaysBefore = null,
+}: MockReservationUnitProps): ReservationUnitType {
+  const reservationUnit: ReservationUnitType = {
+    id: "1",
+    bufferTimeBefore: 60 * 60 * bufferTimeBefore,
+    bufferTimeAfter: 60 * 60 * bufferTimeAfter,
+    maxReservationDuration,
+    minReservationDuration,
+    reservationStartInterval: interval,
+    reservationsMaxDaysBefore,
+    reservationsMinDaysBefore,
+    reservationBegins: addDays(new Date(), -1).toISOString(),
+    reservationEnds: addDays(new Date(), 180).toISOString(),
+  };
+  return reservationUnit;
+}
+
+export function createMockApplicationRound({
+  pk = 1,
+  status,
+  applicationPeriodEnd,
+  applicationPeriodBegin,
+}: {
+  pk?: number;
+  status: ApplicationRoundStatusChoice;
+  applicationPeriodEnd: Date;
+  applicationPeriodBegin: Date;
+}): Readonly<ApplicationRoundFieldsFragment> {
+  return {
+    id: base64encode(`ApplicationRoundNode:${pk}`),
+    pk,
+    nameFi: `ApplicationRound ${pk} FI`,
+    nameSv: `ApplicationRound ${pk} SV`,
+    nameEn: `ApplicationRound ${pk} EN`,
+    status,
+    reservationPeriodBegin: "2024-02-01T00:00:00Z",
+    reservationPeriodEnd: "2025-01-01T00:00:00Z",
+    publicDisplayBegin: "2024-02-01T00:00:00Z",
+    publicDisplayEnd: "2025-01-01T00:00:00Z",
+    applicationPeriodBegin: applicationPeriodBegin.toISOString(),
+    applicationPeriodEnd: applicationPeriodEnd.toISOString(),
+    reservationUnits: [1, 2, 3].map((pk) => ({
+      id: base64encode(`ReservationUnitNode:${pk}`),
+      pk,
+      unit: {
+        id: base64encode(`UnitNode:${pk}`),
+        pk,
+      },
+    })),
+  };
+}
+
+export function createMockReservationUnitType(
+  props: { name: string; pk?: number } | null
+) {
+  if (props == null) {
+    return null;
+  }
+  const { name, pk } = props;
+  return {
+    id: `ReservationUnitTypeNode:${pk ?? 1}`,
+    pk: pk ?? 1,
+    ...generateNameFragment(name),
+  };
+}
+
+export function generateNameFragment(name: string) {
+  return {
+    nameFi: `${name} FI`,
+    nameSv: `${name} SV`,
+    nameEn: `${name} EN`,
   };
 }
