+++ apps/ui/test/test.utils.ts
@@ -1,14 +1,5 @@
-import {
-  ApplicationRoundFieldsFragment,
-  ApplicationRoundStatusChoice,
-  type IsReservableFieldsFragment,
-} from "@/gql/gql-types";
-import { type ReservableMap, type RoundPeriod } from "@/modules/reservable";
 import { fireEvent, render, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { ReservationStartInterval } from "common/gql/gql-types";
-import { base64encode } from "common/src/helpers";
-import { addDays } from "date-fns";
 import { expect } from "vitest";
 
 // Default implementation fakes all timers (expect tick), remove performance from the list
@@ -37,104 +28,8 @@ export const escKeyPressHelper = (): boolean =>
 export const tabKeyPressHelper = (): boolean =>
   fireEvent.keyDown(document, { code: 9, key: "Tab" });
 
-type ReservationUnitType = Omit<
-  IsReservableFieldsFragment,
-  "reservableTimeSpans"
->;
-type MockReservationUnitProps = {
-  bufferTimeBefore?: number;
-  bufferTimeAfter?: number;
-  reservableTimes?: ReservableMap;
-  interval?: ReservationStartInterval;
-  maxReservationDuration?: IsReservableFieldsFragment["maxReservationDuration"];
-  minReservationDuration?: IsReservableFieldsFragment["minReservationDuration"];
-  activeApplicationRounds?: RoundPeriod[];
-  reservationsMinDaysBefore?: number;
-  reservationsMaxDaysBefore?: number | null;
-};
-/// create a mock for IsReservableFragment (not a full reservation unit)
-export function createMockReservationUnit({
-  bufferTimeBefore = 0,
-  bufferTimeAfter = 0,
-  interval = ReservationStartInterval.Interval_15Mins,
-  maxReservationDuration = 0,
-  minReservationDuration = 0,
-  reservationsMinDaysBefore = 0,
-  reservationsMaxDaysBefore = null,
-}: MockReservationUnitProps): ReservationUnitType {
-  const reservationUnit: ReservationUnitType = {
-    id: "1",
-    bufferTimeBefore: 60 * 60 * bufferTimeBefore,
-    bufferTimeAfter: 60 * 60 * bufferTimeAfter,
-    maxReservationDuration,
-    minReservationDuration,
-    reservationStartInterval: interval,
-    reservationsMaxDaysBefore,
-    reservationsMinDaysBefore,
-    reservationBegins: addDays(new Date(), -1).toISOString(),
-    reservationEnds: addDays(new Date(), 180).toISOString(),
-  };
-  return reservationUnit;
-}
-
-export function createMockApplicationRound({
-  pk = 1,
-  status,
-  applicationPeriodEnd,
-  applicationPeriodBegin,
-}: {
-  pk?: number;
-  status: ApplicationRoundStatusChoice;
-  applicationPeriodEnd: Date;
-  applicationPeriodBegin: Date;
-}): Readonly<ApplicationRoundFieldsFragment> {
-  return {
-    id: base64encode(`ApplicationRoundNode:${pk}`),
-    pk,
-    nameFi: `ApplicationRound ${pk} FI`,
-    nameSv: `ApplicationRound ${pk} SV`,
-    nameEn: `ApplicationRound ${pk} EN`,
-    status,
-    reservationPeriodBegin: "2024-02-01T00:00:00Z",
-    reservationPeriodEnd: "2025-01-01T00:00:00Z",
-    publicDisplayBegin: "2024-02-01T00:00:00Z",
-    publicDisplayEnd: "2025-01-01T00:00:00Z",
-    applicationPeriodBegin: applicationPeriodBegin.toISOString(),
-    applicationPeriodEnd: applicationPeriodEnd.toISOString(),
-    reservationUnits: [1, 2, 3].map((pk) => ({
-      id: base64encode(`ReservationUnitNode:${pk}`),
-      pk,
-      unit: {
-        id: base64encode(`UnitNode:${pk}`),
-        pk,
-      },
-    })),
-  };
-}
-
-export function createMockReservationUnitType(
-  props: { name: string; pk?: number } | null
-) {
-  if (props == null) {
-    return null;
-  }
-  const { name, pk } = props;
-  return {
-    id: `ReservationUnitTypeNode:${pk ?? 1}`,
-    pk: pk ?? 1,
-    ...generateNameFragment(name),
-  };
-}
-
-export function generateNameFragment(name: string) {
-  return {
-    nameFi: `${name} FI`,
-    nameSv: `${name} SV`,
-    nameEn: `${name} EN`,
-  };
-}
-
-// TODO write a different version of this that selects Nth option
+// TODO test if we can refactor this not to include any expect calls
+// use throws instead so the error bubbles to the caller
 export async function selectOption(
   view: ReturnType<typeof render> | ReturnType<typeof within>,
   listLabel: RegExp | string,
